feat(data): update ALL test and sample files with partition terminology (DATA-0077 Phase 3d - MASSIVE)
**Test File Updates:**
- Renamed test files: SetRouting* → PartitionRouting*
- Renamed test classes: *SetRouting* → *PartitionRouting*
- DataSetContext.With → EntityContext.Partition
- DataSetContext.Current → EntityContext.Current?.Partition
- All 'string set' parameters → 'string partition'

**Sample File Updates (automatic via sed):**
- Updated 200+ sample files with partition terminology
- Replaced DataSetContext references with EntityContext

**Stats:**
- 11 test files directly updated
- 200+ files auto-updated via batch sed operations

Refs DATA-0077
